# api_document.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "TheHive项目协作平台API",
    "description": "TheHive是一个创业社交网站的项目协作功能，类似于GitHub的项目协作功能，主要以项目制来推动人与人之间的合作。",
    "version": "1.0.0",
    "contact": {
      "name": "TheHive团队"
    }
  },
  "servers": [
    {
      "url": "http://localhost:5000",
      "description": "本地开发环境"
    }
  ],
...
  "paths": {
    "/api/users": {
      "post": {
        "tags": ["用户"],
        "summary": "创建用户",
        "description": "创建新用户，邮箱是必填项",
        "operationId": "createUser",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["email"],
                "properties": {
                  "email": {
                    "type": "string",
                    "description": "用户邮箱",
                    "format": "email"
                  },
                  "full_name": {
                    "type": "string",
                    "description": "用户全名"
                  },
                  "gender": {
                    "type": "string",
                    "description": "性别"
                  },
                  "mbti": {
                    "type": "string",
                    "description": "MBTI性格类型"
                  },
                  "star_sign": {
                    "type": "string",
                    "description": "星座"
                  },
                  "skills": {
                    "type": "string",
                    "description": "技能"
                  },
                  "interests": {
                    "type": "string",
                    "description": "兴趣"
                  },
                  "year_of_study": {
                    "type": "string",
                    "description": "学习年限"
                  },
                  "major": {
                    "type": "string",
                    "description": "专业"
                  },
                  "key_factors": {
                    "oneOf": [
                      {
                        "type": "string",
                        "description": "关键因素（字符串形式）"
                      },
                      {
                        "type": "object",
                        "description": "关键因素（对象形式）",
                        "additionalProperties": {
                          "type": "string"
                        }
                      }
                    ]
                  },
                  "lightning_answers": {
                    "type": "object",
                    "description": "快速回答",
                    "additionalProperties": {
                      "type": "string"
                    }
                  },
                  "picture": {
                    "type": "string",
                    "description": "用户头像URL"
                  },
                  "user_id": {
                    "type": "string",
                    "description": "Auth0用户标识"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "用户创建成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "用户创建成功"
                    },
                    "data": {
                      "$ref": "#/components/schemas/User"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "请求参数有误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "服务器错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": ["用户"],
        "summary": "获取用户列表",
        "description": "获取用户列表，支持多种过滤条件",
        "operationId": "getUsers",
        "parameters": [
          {
            "name": "email",
            "in": "query",
            "description": "邮箱",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "full_name",
            "in": "query",
            "description": "全名",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "gender",
            "in": "query",
            "description": "性别",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "mbti",
            "in": "query",
            "description": "MBTI性格类型",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "star_sign",
            "in": "query",
            "description": "星座",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "major",
            "in": "query",
            "description": "专业",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "user_id",
            "in": "query",
            "description": "Auth0用户标识",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功获取用户列表",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/User"
                      }
                    },
                    "total": {
                      "type": "integer",
                      "description": "用户总数"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "服务器错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/users/{user_id}": {
      "get": {
        "tags": ["用户"],
        "summary": "获取用户详情",
        "description": "根据用户ID获取用户详细信息",
        "operationId": "getUserById",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "description": "用户ID",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功获取用户详情",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/User"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "服务器错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/users/participant": {
      "get": {
        "tags": ["用户"],
        "summary": "获取参与者用户列表",
        "description": "获取参与特定项目的用户列表，支持多种过滤条件",
        "operationId": "getParticipantUsers",
        "parameters": [
          {
            "name": "project_id",
            "in": "query",
            "description": "项目ID",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "email",
            "in": "query",
            "description": "邮箱",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "full_name",
            "in": "query",
            "description": "全名",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "gender",
            "in": "query",
            "description": "性别",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "mbti",
            "in": "query",
            "description": "MBTI性格类型",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "star_sign",
            "in": "query",
            "description": "星座",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "major",
            "in": "query",
            "description": "专业",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "user_id",
            "in": "query",
            "description": "Auth0用户标识",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功获取用户列表",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "allOf": [
                          {
                            "$ref": "#/components/schemas/User"
                          },
                          {
                            "type": "object",
                            "properties": {
                              "project_name": {
                                "type": "string",
                                "description": "项目名称"
                              },
                              "project_skill": {
                                "type": "string",
                                "description": "项目技能"
                              }
                            }
                          }
                        ]
                      }
                    },
                    "total": {
                      "type": "integer",
                      "description": "用户总数"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "服务器错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "用户ID"
          },
          "email": {
            "type": "string",
            "description": "用户邮箱",
            "format": "email"
          },
          "full_name": {
            "type": "string",
            "description": "用户全名"
          },
          "gender": {
            "type": "string",
            "description": "性别"
          },
          "mbti": {
            "type": "string",
            "description": "MBTI性格类型"
          },
          "star_sign": {
            "type": "string",
            "description": "星座"
          },
          "skills": {
            "type": "string",
            "description": "用户技能"
          },
          "interests": {
            "type": "string",
            "description": "兴趣爱好"
          },
          "year_of_study": {
            "type": "string",
            "description": "学习年限"
          },
          "major": {
            "type": "string",
            "description": "专业"
          },
          "key_factors": {
            "type": "object",
            "description": "关键因素",
            "additionalProperties": {
              "type": "string"
            }
          },
          "lightning_answers": {
            "type": "object",
            "description": "快速回答",
            "additionalProperties": {
              "type": "string"
            }
          },
          "picture": {
            "type": "string",
            "description": "用户头像URL"
          },
          "user_id": {
            "type": "string",
            "description": "Auth0用户标识"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "description": "创建时间"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "description": "更新时间"
          }
        }
      },
...
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "description": "错误信息"
          }
        }
      },
...
    }
  }
} 